Create RuntimeDashboard
This is the runtime dashboard I've been working on. It required the installation of some packages that are not on the python version stored on the shared drive - what this means is that for the team to use this tool, we will need them to grab the new PythonTT_Summary folder from the S: drive (like normal), the new TAIPAN v2.3 file from the S: drive (like normal) and ALSO require them to copy the "python-3.7.4.amd64" folder from the S: drive to their local drive once I have moved my updated version of this folder to the shared drive. I will send an email out shortly to ellaborate.